issue #451
Switch order of onboarding screens so "Totally anonymous" comes before "What the app does not do" #451
