+++ src/screens/onboarding/views/WhatItsNot.tsx
@@ -1,16 +1,16 @@
-import React, {useCallback} from 'react';
-import {ScrollView, StyleSheet} from 'react-native';
-import {Box, Text, OnboardingHeader} from 'components';
-import {useI18n} from '@shopify/react-i18n';
-import {useNavigation} from '@react-navigation/native';
-import {SafeAreaView} from 'react-native-safe-area-context';
+import React, { useCallback } from 'react';
+import { ScrollView, StyleSheet } from 'react-native';
+import { Box, Text, OnboardingHeader } from 'components';
+import { useI18n } from '@shopify/react-i18n';
+import { useNavigation } from '@react-navigation/native';
+import { SafeAreaView } from 'react-native-safe-area-context';
 
-import {BackButton, NextButton, StepText} from '../components';
+import { BackButton, NextButton, StepText } from '../components';
 
 export const WhatItsNot = () => {
   const [i18n] = useI18n();
   const navigation = useNavigation();
-  const onNext = useCallback(() => navigation.navigate('OnboardingAnonymous'), [navigation]);
+  const onNext = useCallback(() => navigation.navigate('OnboardingHowItWorks'), [navigation]);
 
   return (
     <Box flex={1} backgroundColor="overlayBackground">
@@ -39,7 +39,7 @@ export const WhatItsNot = () => {
             </Box>
           </ScrollView>
           <Box height={5} maxHeight={2} borderTopWidth={2} borderTopColor="gray5" />
-          <StepText index={2} />
+          <StepText index={3} />
           <NextButton onNext={onNext} />
         </Box>
       </SafeAreaView>
